+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support net core plugin.
 ### Added
 - Add `CodeAnalysisCodeDom` to support net code.
+### Tests
+- Test `GistGithubUtils` download string.
+- Add `Newtonsoft.Json` to support `GistGithubUtils` download string.
 
 ## [1.0.6] / 2024-01-27
 ### Features